+++ yash-env/CHANGELOG.md
@@ -9,6 +9,13 @@ Terminology: A _public dependency_ is one that’s exposed through this crate’
 public API (e.g., re-exported types).
 A _private dependency_ is used internally and not visible to downstream users.
 
+## [0.13.0] - Unreleased
+
+### Changed
+
+- The `Stat` struct has been moved from the `system` module to the
+  `system::virtual` module.
+
 ## [0.12.1] - 2026-02-06
 
 ### Fixed
@@ -862,6 +869,7 @@ A _private dependency_ is used internally and not visible to downstream users.
 
 - Initial implementation of the `yash-env` crate
 
+[0.13.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.13.0
 [0.12.1]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.12.1
 [0.12.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.12.0
 [0.11.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.11.0